# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_api.recentPRsEntry:
    properties:
      author:
        type: string
      merged_at:
        type: string
      pr_id:
        type: integer
      title:
        type: string
      url:
        type: string
    type: object
  internal_api.recentPRsResponse:
    properties:
      entries:
        items:
          $ref: '#/definitions/internal_api.recentPRsEntry'
        type: array
      values_file:
        type: string
      view_all_url:
        type: string
    type: object
...
  internal_api.setAddonValuesRequest:
    properties:
      refresh_from_upstream:
        description: |-
          RefreshFromUpstream, when true, regenerates the global values file
          from the chart's upstream values.yaml at the catalog-pinned version.
          The audit event for this path is `values_refreshed_from_upstream`
          (overriding the default `values_set`) so audit consumers can
          distinguish manual edits from upstream refreshes.
        type: boolean
      values:
        description: |-
          Values is the full YAML file content. Sharko commits this verbatim
          as the new global default values for the addon. Ignored when
          RefreshFromUpstream is true.
        type: string
    type: object
  internal_api.setClusterAddonValuesRequest:
    properties:
      values:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Addon management server for Kubernetes clusters, built on ArgoCD.
  title: Sharko API
  version: 1.0.0
paths:
...
  /addons/{name}/values:
    get:
      description: Returns the default Helm values file for a specific addon
      parameters:
      - description: Addon name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Addon values
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get addon values
      tags:
      - addons
    put:
      consumes:
      - application/json
      description: 'Replaces the global default Helm values YAML for an addon and
        opens a PR. Tier 2 (configuration) — prefers the caller''s personal GitHub
        PAT for proper Git authorship; otherwise falls back to the service token with
        a `Co-authored-by:` trailer and surfaces `attribution_warning: "no_per_user_pat"`
        in the response.'
      parameters:
      - description: Addon name
        in: path
        name: name
        required: true
        type: string
      - description: Full YAML payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.setAddonValuesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: PR created (or merged if auto-merge is on)
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid YAML or missing field
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Addon not found
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Git or ArgoCD failure
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Edit global addon values
      tags:
      - addons
...
  /addons/{name}/values/recent-prs:
    get:
      description: Returns up to `limit` (default 5, max 20) recently-merged PRs that
        modified the addon's global values file. Backed by a 5-minute in-memory cache.
      parameters:
      - description: Addon name
        in: path
        name: name
        required: true
        type: string
      - description: Maximum entries (default 5, max 20)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.recentPRsResponse'
      security:
      - BearerAuth: []
      summary: Recent PRs that touched an addon's global values
      tags:
      - addons
...
  /clusters/{cluster}/addons/{name}/values:
    get:
      description: Returns the YAML for a single addon's section in a cluster's overrides
        file, plus an optional JSON Schema. `current_overrides` is empty when no overrides
        are configured for this addon yet.
      parameters:
      - description: Cluster name
        in: path
        name: cluster
        required: true
        type: string
      - description: Addon name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Current overrides + optional schema
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Git failure
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get per-cluster addon overrides + schema
      tags:
      - clusters
    put:
      consumes:
      - application/json
      description: Replaces the per-cluster overrides YAML for one addon on one cluster
        and opens a PR. Pass an empty `values` string to remove the addon's overrides
        entirely. Tier 2 (configuration) — same attribution semantics as the global
        values endpoint.
      parameters:
      - description: Cluster name
        in: path
        name: cluster
        required: true
        type: string
      - description: Addon name
        in: path
        name: name
        required: true
        type: string
      - description: Addon-section YAML payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.setClusterAddonValuesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: PR created (or merged if auto-merge is on)
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid YAML or missing field
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Git or ArgoCD failure
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Edit per-cluster addon overrides
      tags:
      - clusters
  /clusters/{cluster}/addons/{name}/values/recent-prs:
    get:
      description: Returns up to `limit` (default 5, max 20) recently-merged PRs that
        modified the cluster overrides file, filtered to the addon name.
      parameters:
      - description: Cluster name
        in: path
        name: cluster
        required: true
        type: string
      - description: Addon name
        in: path
        name: name
        required: true
        type: string
      - description: Maximum entries (default 5, max 20)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.recentPRsResponse'
      security:
      - BearerAuth: []
      summary: Recent PRs that touched per-cluster addon overrides
      tags:
      - clusters
...
  /clusters/{name}/values:
    get:
      description: Returns the Helm values file for a specific cluster
      parameters:
      - description: Cluster name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Cluster values
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get cluster values
      tags:
      - clusters
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"